TICKET-144: split ReplayNode DOM overlay into ReplayOverlayNode
Extract DOM overlay concerns (letterbox bars, flash transition, REPLAY
label, self-KO text with per-letter bobbing) into ReplayOverlayNode.
ReplayNode now only handles replay playback advancement, particle VFX,
camera shake, hit impact bursts, trail emission, and sound.

Also fix tests to use vi.mock instead of jest.mock for vitest compat.
